@@ -19,71 +19,61 @@
     ]
   },
   "dependencies": {
-    "@emotion/cache": "^11.4.0",
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@material-ui/core": "5.0.0-alpha.34",
-    "@material-ui/icons": "5.0.0-alpha.34",
-    "@material-ui/lab": "5.0.0-alpha.34",
-    "@material-ui/system": "5.0.0-alpha.34",
-    "@zeit/next-css": "^1.0.1",
-    "axios": "^0.21.1",
+    "@emotion/cache": "^11.6.0",
+    "@emotion/react": "^11.6.0",
+    "@emotion/server": "^11.4.0",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.0",
+    "@mui/lab": "^5.0.0-alpha.55",
+    "@mui/material": "^5.1.1",
+    "@mui/styles": "^5.1.1",
+    "@mui/system": "^5.1.1",
+    "axios": "^0.24.0",
     "d3": "^5.16.0",
-    "date-fns": "^2.23.0",
+    "date-fns": "^2.26.0",
     "formik": "^2.2.9",
-    "history": "^5.0.1",
-    "i18next": "^20.4.0",
+    "i18next": "^21.5.2",
     "jquery": "^3.6.0",
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
-    "next": "^11.1.0",
+    "next": "^12.0.5-canary.6",
     "nprogress": "^0.2.0",
-    "prop-types": "^15.7.2",
     "react": "^17.0.2",
-    "react-apexcharts": "^1.3.9",
-    "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^17.0.2",
-    "react-draft-wysiwyg": "^1.14.7",
-    "react-dropzone": "^11.3.4",
-    "react-helmet-async": "^1.0.9",
-    "react-hot-toast": "^2.1.0",
-    "react-i18next": "^11.11.4",
-    "react-markdown": "^7.0.0",
-    "react-modal-image": "^2.5.0",
+    "react-dropzone": "^11.4.2",
+    "react-helmet-async": "^1.1.2",
+    "react-hot-toast": "^2.1.1",
+    "react-i18next": "^11.14.2",
     "react-perfect-scrollbar": "^1.5.8",
-    "react-quill": "^1.3.5",
-    "react-redux": "^7.2.4",
-    "react-router": "^6.0.0-beta.0",
-    "react-syntax-highlighter": "^15.4.4",
-    "stylis-plugin-rtl": "^2.1.0",
-    "yup": "^0.32.9"
+    "yup": "^0.32.11"
   },
   "devDependencies": {
     "@types/d3": "^5.16.0",
-    "@types/jquery": "^3.5.6",
-    "@types/lodash": "^4.14.172",
+    "@types/jquery": "^3.5.8",
+    "@types/lodash": "^4.14.177",
     "@types/nprogress": "^0.2.0",
-    "@types/react": "^17.0.19",
-    "@types/react-beautiful-dnd": "^13.1.1",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-draft-wysiwyg": "^1.13.3",
-    "@types/react-router-dom": "^5.1.8",
+    "@types/react": "^17.0.35",
+    "@types/react-dom": "^17.0.11",
     "@types/yup": "^0.29.13",
-    "@typescript-eslint/eslint-plugin": "^4.30.0",
-    "eslint": "^7.32.0",
-    "eslint-config-next": "11.1.0",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "eslint": "^8.2.0",
+    "eslint-config-next": "^12.0.5-canary.6",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "husky": "^7.0.2",
-    "lint-staged": "^11.1.2",
-    "prettier": "^2.3.2",
-    "prettier-plugin-packagejson": "^2.2.11",
-    "prettier-plugin-sh": "^0.7.1",
-    "typescript": "^4.2.4",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "husky": "^7.0.4",
+    "lint-staged": "^12.0.3",
+    "prettier": "^2.4.1",
+    "prettier-plugin-packagejson": "^2.2.15",
+    "prettier-plugin-sh": "^0.8.0",
+    "typescript": "^4.4.4",
     "yarn-deduplicate": "^3.1.0"
+  },
+  "engines": {
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0",
+    "yarn": "^1.22.0"
   }
 }
