@@ -1,47 +1,51 @@
 import React, { FC } from "react";
-import Card from "@material-ui/core/Card";
-import CardActionArea from "@material-ui/core/CardActionArea";
-import CardActions from "@material-ui/core/CardActions";
-import CardContent from "@material-ui/core/CardContent";
-import CardMedia from "@material-ui/core/CardMedia";
-import Button from "@material-ui/core/Button";
-import Typography from "@material-ui/core/Typography";
-import { createStyles, IconButton, Theme, makeStyles } from "@material-ui/core";
-import Bookmark from "../Bookmark";
+import {
+  Card,
+  CardActionArea,
+  CardContent,
+  CardMedia,
+  Typography,
+} from "@mui/material";
+import { makeStyles } from "@mui/styles";
+import Link from "next/link";
 
 const useStyles = makeStyles({
   root: {
-    maxWidth: 345,
-  },
-  media: {
-    height: 140,
+    width: 300,
+    height: 300,
+    padding: "10px 10px 10px 10px",
   },
 });
 
 interface PortalItemProps {
-  data: any;
+  id: string;
+  image: string;
+  title: string;
 }
 
-const PortalItem: FC<PortalItemProps> = ({ data }) => {
+const PortalItem: FC<PortalItemProps> = ({ id, image, title }) => {
   const classes = useStyles();
-  console.log("PortalItem: data =", data);
+  // prettier-ignore
+  console.log("PortalItem: { image, title } =", image, title);
 
   return (
     <Card className={classes.root}>
       <CardActionArea>
-        <CardMedia className={classes.media} image={data.img} title="" />
+        <CardMedia component="img" height={250} alt="" image={image} title="" />
         <CardContent>
-          <Typography gutterBottom variant="h6" component="h2">
-            {data.title}
-          </Typography>
-          <Typography variant="body2" color="textSecondary" component="p">
-            {data.vis.description}
+          <Link href={{ pathname: "/page", query: { id: id } }} passHref={true}>
+            {" "}
+          </Link>
+
+          <Typography gutterBottom variant="subtitle1">
+            {title}
           </Typography>
         </CardContent>
       </CardActionArea>
-      <CardActions>
+
+      {/* <CardActions>
         <Bookmark pageId="xxxx" />
-      </CardActions>
+      </CardActions> */}
     </Card>
   );
 };
