@@ -1,7 +1,3 @@
-/**
- * CSR
- */
-
 import { ReactElement, useCallback, useEffect, useState } from "react";
 import { Helmet } from "react-helmet-async";
 import {
@@ -12,19 +8,23 @@ import {
   Container,
   Grid,
   Card,
-} from "@material-ui/core";
-import { makeStyles } from "@material-ui/core/styles";
-import { blue } from "@material-ui/core/colors";
+  Fade,
+  CircularProgress,
+} from "@mui/material";
+import { makeStyles } from "@mui/styles";
+import { blue } from "@mui/material/colors";
 import axios from "axios";
-import InsertChartIcon from "@material-ui/icons/InsertChart";
+import InsertChartIcon from "@mui/icons-material/InsertChart";
 import { useRouter } from "next/router";
 
 import useSettings from "src/hooks/useSettings";
 import { visFactory } from "src/lib/vis/vis-factory";
-import useAuth from "src/hooks/useAuth";
 import Bookmark from "src/components/Bookmark";
-import { apiService } from "src/utils/apiService";
+import { apiService } from "src/utils/ApiService";
 import DashboardLayout from "src/components/dashboard-layout/DashboardLayout";
+import { IData } from "src/models/IData";
+import { ILink } from "src/models/ILink";
+import { getLinks } from "src/utils/LinkService";
 
 const API = {
   API_PY: process.env.NEXT_PUBLIC_API_PY,
@@ -39,16 +39,6 @@ const useStyles = makeStyles((theme) => ({
     height: 0,
     paddingTop: "56.25%", // 16:9
   },
-  expand: {
-    transform: "rotate(0deg)",
-    marginLeft: "auto",
-    transition: theme.transitions.create("transform", {
-      duration: theme.transitions.duration.shortest,
-    }),
-  },
-  expandOpen: {
-    transform: "rotate(180deg)",
-  },
   avatar: {
     backgroundColor: blue[500],
   },
@@ -58,53 +48,67 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 const PropagatedPage = () => {
-  // const { pageId } = useParams(); // 60ae9fce8839aa3ae916e217
-  // console.log("PropagatedPage: pageId = ", pageId);
-  // const { user } = useAuth();
-
   const { settings } = useSettings();
   const classes = useStyles();
   const router = useRouter();
   const pageId =
     typeof router.query.id === "string" ? router.query.id : undefined;
   const [title, setTitle] = useState<string>("");
+  const [loading, setLoading] = useState(false);
 
   const fetchOntoPage = useCallback(async () => {
     if (!pageId) {
       return;
     }
-    const page = await apiService.get<any>(`/template/page/${pageId}`);
-    console.log("PropagatedPage: page = ", page);
 
-    setTitle(page?.title);
+    try {
+      setLoading(true);
 
-    const dataForVisFunction = await Promise.all(
-      page?.data?.map(async (d: any) => {
-        const endpoint = `${API[d.urlCode]}${d.endpoint}`;
-        const values = (await axios.get(endpoint)).data;
-        const { description } = d;
-        return { endpoint, values, description };
-      }),
-    );
+      const page = await apiService.get(`/template/page/${pageId}`);
+      // eslint-disable-next-line no-console -- VIS developers need....
+      console.log("[TEMPLATE] VIS Function = ", page.vis);
+      // eslint-disable-next-line no-console -- VIS developers need....
+      console.log("[TEMPLATE] Data = ", page.data);
 
-    const links = page?.pageIds?.map((d: any) => {
-      console.log(d);
-      return `page/${d}`;
-    });
+      setTitle(page?.title);
 
-    console.log("PropagatedPage: dataForVisFunction = ", dataForVisFunction);
+      // fetch data streams
+      const data = await Promise.all(
+        page?.data?.map(async (d: any) => {
+          const endpoint = `${API[d.urlCode]}${d.endpoint}`;
+          const values = (await axios.get(endpoint)).data;
+          const { id, description } = d;
+          return { id, endpoint, values, description } as IData;
+        }),
+      );
 
-    visFactory(page?.vis?.function, {
-      chartElement: "charts", // ref.current,
-      data: dataForVisFunction,
-      links,
-    });
+      // fetch links
+      const links: ILink[] = await Promise.all(
+        page?.data?.map(async (d: any) => {
+          const links: ILink[] = await getLinks(d.id);
+          return links;
+        }),
+      );
+
+      // eslint-disable-next-line no-console -- VIS developers need....
+      console.log("[TEMPLATE] fetched data = ", data);
+      console.log("[TEMPLATE] fetched links = ", links);
+
+      visFactory(page?.vis?.function, {
+        chartElement: "charts",
+        data: data,
+        links: links,
+      });
+
+      setLoading(false);
+    } catch (err) {
+      // prettier-ignore
+      console.error(`PropagatedPage: Fetching data error = ${err}`);
+      setLoading(false);
+    }
   }, [pageId]);
-  // if pageId changes, useEffect will run again
-  // if you want to run only once, just leave array empty []
 
   useEffect(() => {
-    console.log("PropagatedPage: useEffect:");
     fetchOntoPage();
   }, [fetchOntoPage]);
 
@@ -126,12 +130,13 @@ const PropagatedPage = () => {
             <Grid item xs={12}>
               <Card>
                 <CardHeader
-                  action={
-                    // <IconButton aria-label="settings">
-                    //   { user?.id ? (<MoreVertIcon />) : (<TimelineIcon />)}
-                    // </IconButton>
-                    <Bookmark pageId={pageId} />
-                  }
+                  // TODO:
+                  // action={
+                  //   <IconButton aria-label="settings">
+                  //     { user?.id ? (<MoreVertIcon />) : (<TimelineIcon />)}
+                  //   </IconButton>
+                  //   <Bookmark pageId={pageId} />
+                  // }
                   avatar={
                     <Avatar className={classes.avatar}>
                       <InsertChartIcon />
@@ -142,6 +147,19 @@ const PropagatedPage = () => {
                 />
 
                 <CardContent sx={{ pt: "8px" }}>
+                  {loading && (
+                    <Box sx={{ height: 40 }}>
+                      <Fade
+                        in={loading}
+                        style={{
+                          transitionDelay: loading ? "800ms" : "0ms",
+                        }}
+                        unmountOnExit
+                      >
+                        <CircularProgress />
+                      </Fade>
+                    </Box>
+                  )}
                   <div id="charts" />
                 </CardContent>
               </Card>
