@@ -1,10 +1,6 @@
-/* eslint-disable arrow-body-style */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-
 import React, { ReactElement, useCallback, useEffect, useState } from "react";
 import { Helmet } from "react-helmet-async";
-import useMediaQuery from "@material-ui/core/useMediaQuery";
-import { useTheme, makeStyles } from "@material-ui/core/styles";
+import { useTheme, makeStyles } from "@mui/styles";
 import {
   Avatar,
   Box,
@@ -13,12 +9,11 @@ import {
   Container,
   Grid,
   Card,
-} from "@material-ui/core";
-import { blue } from "@material-ui/core/colors";
-import BookmarksIcon from "@material-ui/icons/Bookmarks";
-
+} from "@mui/material";
+import { blue } from "@mui/material/colors";
+import BookmarksIcon from "@mui/icons-material/Bookmarks";
 import useSettings from "src/hooks/useSettings";
-import { apiService } from "src/utils/apiService";
+import { apiService } from "src/utils/ApiService";
 import DashboardLayout from "src/components/dashboard-layout/DashboardLayout";
 import AuthGuard from "src/components/auth/guards/AuthGuard";
 import PortalView from "src/components/portal/PortalView";
@@ -31,16 +26,6 @@ const useStyles = makeStyles((theme) => ({
   media: {
     height: 0,
     paddingTop: "56.25%", // 16:9
-  },
-  expand: {
-    transform: "rotate(0deg)",
-    marginLeft: "auto",
-    transition: theme.transitions.create("transform", {
-      duration: theme.transitions.duration.shortest,
-    }),
-  },
-  expandOpen: {
-    transform: "rotate(180deg)",
   },
   avatar: {
     backgroundColor: blue[500],
@@ -57,17 +42,16 @@ const MyPortal = () => {
   const { settings } = useSettings();
 
   const [portalData, setPortalData] = useState<any[]>(mockPortalData);
+  const url = `/template/pages/example`; // TODO: Set correct URL once available
 
   const fetchPortalPages = useCallback(async () => {
     try {
-      // TODO: Set correct URL once available
-      // const url = `/template/pages/example/plot/`;
-      // const res = (await apiService.get<any>(url))?.data;
-      // console.log("MyPortal: fetched data = ", res);
-      // setPortalData(res);
+      const res = (await apiService.get(url))?.data;
+      console.log("MyPortal: data = ", res);
+      setPortalData(res);
     } catch (err) {
       // prettier-ignore
-      console.error(`PageListTemplate: Fetching API, error = ${err}`);
+      console.error(`MyPortal: Fetching ${url}, error = ${err}`);
     }
   }, []);
 
@@ -92,9 +76,13 @@ const MyPortal = () => {
             <Grid item xs={12}>
               <Card>
                 <CardHeader
-                  avatar={<BookmarksIcon />}
-                  title="My Bookmarks"
-                  subheader=""
+                  avatar={
+                    <Avatar className={classes.avatar}>
+                      <BookmarksIcon />
+                    </Avatar>
+                  }
+                  title="Portal"
+                  subheader="List of my bookmarked pages"
                 />
 
                 <CardContent sx={{ pt: "8px" }}>
@@ -111,9 +99,9 @@ const MyPortal = () => {
 
 MyPortal.getLayout = function getLayout(page: ReactElement) {
   return (
-    // <AuthGuard>
-    <DashboardLayout>{page}</DashboardLayout>
-    // </AuthGuard>
+    <AuthGuard>
+      <DashboardLayout>{page}</DashboardLayout>
+    </AuthGuard>
   );
 };
 
