@@ -1,23 +1,33 @@
 import { SimpleBarChart } from "./simple-bar-chart";
 import { SimpleLineChart } from "./simple-line-chart";
-import { HealthBoardOverview } from "./health-board-overview";
-import { CouncilOverview } from "./council-overview";
-import { CountryOverview } from "./country-overview";
-import { CountryOverviewNew } from "./country-overview-new";
 import { ChordDiagram } from "./chord-diagram";
 import { Matrix } from "./matrix";
 import { SuperimposedPercentiles } from "./superimposed-percentiles";
 import { StackedBarChartWith6Places } from "./stacked-bar-chart-6-places";
 import { StackedAreaChart } from "./stacked-area-chart";
 import { StackedBarChart } from "./stacked-bar-chart";
+import { MirroredStackedAreaChart } from "./mirrored-stacked-area-chart";
+import { MirroredStackedBarChart } from "./mirrored-stacked-bar-chart";
+import { SensitivityStackedBarChart } from "./sensitivity-stacked-bar-chart";
+import { RiskMonitoring } from "./risk-monitoring";
+import { UncertaintySampleAndMean } from "./uncertainty-sample-and-mean";
+import { UncertaintyClusterSampleAndMean } from "./uncertainty-cluster-sample-and-mean";
+
+// Dashboards
+import { CouncilOverview } from "./dashboards/council-overview";
+import { CountryOverview } from "./dashboards/country-overview";
+import { HealthBoardOverview } from "./dashboards/health-board-overview";
+import { CountryOverviewNew } from "./dashboards/country-overview-new";
+import { VaccineOverview } from "./dashboards/vaccineOverview";
+import { DashboardUK } from "./dashboards/dashboard-UK";
+import { DashboardTian } from "./dashboards/dashboard-tian";
+import { DashboardLowerTierLocalAuthority } from "./dashboards/dashboard-lowerTierLocalAuthority";
+import { DashboardMSOA } from "./dashboards/dashboard-msoa";
+import { DashboardRegion } from "./dashboards/dashboard-region"
 
 export const visFactory = (type, args) => {
   if (type === "SimpleBarChart") return new SimpleBarChart(args);
   if (type === "SimpleLineChart") return new SimpleLineChart(args);
-  if (type === "HealthBoardOverview") return new HealthBoardOverview(args);
-  if (type === "CouncilOverview") return new CouncilOverview(args);
-  if (type === "CountryOverview") return new CountryOverview(args);
-  if (type === "CountryOverviewNew") return new CountryOverviewNew(args);
   if (type === "ChordDiagram") return new ChordDiagram(args);
   if (type === "Matrix") return new Matrix(args);
   if (type === "SuperimposedPercentiles")
@@ -26,6 +36,29 @@ export const visFactory = (type, args) => {
     return new StackedBarChartWith6Places(args);
   if (type === "StackedAreaChart") return new StackedAreaChart(args);
   if (type === "StackedBarChart") return new StackedBarChart(args);
+  if (type === "MirroredStackedAreaChart")
+    return new MirroredStackedAreaChart(args);
+  if (type === "MirroredStackedBarChart")
+    return new MirroredStackedBarChart(args);
+  if (type === "SensitivityStackedBarChart")
+    return new SensitivityStackedBarChart(args);
+  if (type === "RiskMonitoring") return new RiskMonitoring(args);
+  if (type == "UncertaintySampleAndMean")
+    return new UncertaintySampleAndMean(args);
+  if (type == "UncertaintyClusterSampleAndMean")
+    return new UncertaintyClusterSampleAndMean(args);
+
+  // Dashboards
+  if (type === "CouncilOverview") return new CouncilOverview(args);
+  if (type === "HealthBoardOverview") return new HealthBoardOverview(args);
+  if (type === "CountryOverview") return new CountryOverview(args);
+  // if (type === "CountryOverviewNew") return new CountryOverviewNew(args);
+  if (type === "CountryOverviewNew") return new DashboardTian(args);
+  // if (type === "VaccineOverview") return new VaccineOverview(args);
+  if (type === "DashboardUK") return new DashboardUK(args);
+  // if (type === "VaccineOverview") return new DashboardLowerTierLocalAuthority(args);
+  // if (type === "VaccineOverview") return new DashboardMSOA(args);
+  // if (type === "VaccineOverview") return new DashboardRegion(args);
 
   return null;
 };
