@@ -1,15 +1,12 @@
-/* eslint-disable arrow-body-style */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-
 import { ReactElement } from "react";
 import { Helmet } from "react-helmet-async";
 import { useRouter } from "next/router";
 import _ from "lodash";
 import DashboardLayout from "src/components/dashboard-layout/DashboardLayout";
 import UnderConstruction from "src/components/errors/UnderConstruction";
-import { GetStaticProps, GetStaticPaths, NextPage } from "next";
+import { GetStaticProps, GetStaticPaths } from "next";
 
-const MyPortal: NextPage = () => {
+const MyPortal = () => {
   const router = useRouter();
   const country = `${router.query.country}`;
 
